Minor fixes
+ changed int datatypes and params for IDs to long
~ fixed PathVariables for Dog and User
~ fixed capitalization errors
- deleted CountryRepo

+++ Backend/src/main/java/com/it192/backend/repository/DogRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.it192.backend.model.Dog;
 
 @Repository
-public interface DogRepository extends CrudRepository <Dog, Integer> {
+public interface DogRepository extends CrudRepository <Dog, Long> {
 }
